fixes search bar flickering
when the search starts, for a single frame it was setting noDataFound to true when searchLoading was still false. Due to the useEffect that actually starts the search, this searchLoading flag was converted to true after a single frame. This was causing a little flickering. I disbanded this useEffect, and made these flags changing together in searchHandle function